requirements.txt
django>=4.1.2
pytz==2017.2
django-materializecss-form>=1.1.17
django-allauth>=0.44.0
google-api-python-client>=2.64.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.3
python-dotenv>=0.21.0
